P1: Job 12 - transformations, AlterSafe, declared renames, enum lookup sync (1.0.6)
The fourth-pass taxonomy (D-233), implemented:

AlterSafe - a new classification for provably-lossless widenings: same-family
length growth (to bigger n or MAX), the integer ladder upward, decimal growth
that shrinks neither integer digits nor scale, fractional-seconds precision
up, datetime->datetime2, NOT NULL relaxing to NULL. Joins the auto-apply set
under the same contract as AddSafe ("provably cannot lose data"); its
ROLLBACK is @Force-gated because the inverse is a narrowing. Narrowings,
cross-family retypes, and NULL->NOT NULL tightenings stay script-only.

Declared transformations - schema comparison cannot recover INTENT, so it is
declared with self-retiring attributes that fire only while their SOURCE
still exists: [RenamedFromColumn] (collapses the Add+Remove pair into ONE
Rename; sp_rename script-only because code on the old package still reads
the old name), [RenamedFromTable] (the new table boots empty; the composed
script moves rows, count-verifies, drops), [MigratedFromColumn] (column->
table promotion: INSERT with the estate's GUID-string PK + audit stamps,
count-verify THROW, only then the source column drops), [FlattenedFromTable]
(child rows serialize into the bridge column via FOR JSON, count-verify,
drop table). Every script: COPY -> VERIFY -> THROW on mismatch -> guarded
destructive step. Transformations never auto-apply in any mode, permanently.

Tier 1 teaches Tier 2 (operator requirement): an undeclared same-shape
Add+Remove pair annotates BOTH reasons with the [RenamedFromColumn] that
would make it whole, and every Remove reason names [MigratedFromColumn] -
fix it in code or finish the scaffold by hand.

Enum lookup sync - the standing landmine (taxonomy #13): seeding fired only
when the table was MISSING, so a new enum member never got its lookup row
and every later insert failed its FK. Now missing members INSERT additively
at boot; renames stay untouched (a display concern, not an FK one).

Proven: 104/104 including two new real-SQL round trips (int->bigint widens
automatically with the row's value intact; a gutted enum table heals on the
next boot) and 6/6 mutations reddening disjoint tests - narrowing-as-widen,
tightening allowed, AlterSafe dropped from the auto set, rename not
consuming, promotion without its THROW, rollback without its Force gate.

+++ Nostreets.Orm.EF.Core.Test/SchemaMigrationTests.cs
@@ -210,9 +210,12 @@ namespace Nostreets.Orm.EF.Core.Test
                     Pk,
                     Model("NewNullable", "nvarchar(max)"),
                     Model("NewRequired", "int", nullable: false),
-                    Model("Retyped", "bigint")
+                    // NARROWING on purpose: bigint -> int can lose data, so it must stay OUT of
+                    // the auto set. (The widening direction is AlterSafe and IS in the set - covered
+                    // in SchemaTransformTests.)
+                    Model("Retyped", "int")
                 },
-                new[] { LivePk, Live("Retyped", "int"), Live("Orphan", "bit") });
+                new[] { LivePk, Live("Retyped", "bigint"), Live("Orphan", "bit") });
 
             SchemaDriftAnalyzer.AdditiveSafe(drifts).Should().ContainSingle()
                 .Which.ColumnName.Should().Be("NewNullable");
